Update AutoInsurance_Regression_FinalProject.R
Code has a detailed overview analysis of an automobile insurance company.

Any kind of damage caused to the automobiles be it a corporate vehicle, personal one or for people who are likely to get into accidents but have limited financial resource, then paying from own pocket is very painful as it is very expensive. This is where insurance company comes into play. People tend to pay premiums to automobile insurance companies; in turn they pay for the damages. There exists a contract between the customer and the auto insurance company, that protects the customer against any kind of damages caused due to accidents or theft to the customer’s vehicles. In exchange of customers paying for a premium plan and hence the company agrees to pay for the customer’s losses as outlines in their contract or policy.